Move merkle pkg to internal dir
