+++ internal/merkle/clock/clock.go
