+++ internal/merkle/clock/errors.go
