+++ internal/merkle/crdt/errors.go
