+++ internal/merkle/crdt/merklecrdt.go
