+++ internal/merkle/doc.go
